mod_attendance: version 2016082900 (3.1.1.1), commit 08cc9e88c5c

+++ mod/attendance/sessions.php
@@ -59,7 +59,6 @@ $PAGE->set_url($att->url_sessions(array('action' => $pageparams->action)));
 $PAGE->set_title($course->shortname. ": ".$att->name);
 $PAGE->set_heading($course->fullname);
 $PAGE->set_cacheable(true);
-$PAGE->set_button($OUTPUT->update_module_button($cm->id, 'attendance'));
 $PAGE->navbar->add($att->name);
 
 $currenttab = attendance_tabs::TAB_ADD;
@@ -133,15 +132,25 @@ switch ($att->pageparams->action) {
         echo $OUTPUT->footer();
         exit;
     case mod_attendance_sessions_page_params::ACTION_DELETE_SELECTED:
+    case mod_attendance_sessions_page_params::ACTION_CREATE_CAL_EVENTS:
+    case mod_attendance_sessions_page_params::ACTION_DELETE_CAL_EVENTS:
         $confirm    = optional_param('confirm', null, PARAM_INT);
+        $message = get_string('deletecheckfull', '', get_string('session', 'attendance'));
 
         if (isset($confirm) && confirm_sesskey()) {
             $sessionsids = required_param('sessionsids', PARAM_ALPHANUMEXT);
             $sessionsids = explode('_', $sessionsids);
-
-            $att->delete_sessions($sessionsids);
-            attendance_update_users_grade($att);
-            redirect($att->url_manage(), get_string('sessiondeleted', 'attendance'));
+            if ($att->pageparams->action == mod_attendance_sessions_page_params::ACTION_DELETE_SELECTED) {
+                $att->delete_sessions($sessionsids);
+                attendance_update_users_grade($att);
+                redirect($att->url_manage(), get_string('sessiondeleted', 'attendance'));
+            } else if ($att->pageparams->action == mod_attendance_sessions_page_params::ACTION_CREATE_CAL_EVENTS) {
+                attendance_create_calendar_events($sessionsids);
+                redirect($att->url_manage(), get_string('createcheckcalevents', 'attendance'));
+            } else if ($att->pageparams->action == mod_attendance_sessions_page_params::ACTION_DELETE_CAL_EVENTS) {
+                attendance_delete_calendar_events($sessionsids);
+                redirect($att->url_manage(), get_string('deletecheckcalevents', 'attendance'));
+            }
         }
         $sessid = optional_param_array('sessid', '', PARAM_SEQUENCE);
         if (empty($sessid)) {
@@ -149,7 +158,12 @@ switch ($att->pageparams->action) {
         }
         $sessionsinfo = $att->get_sessions_info($sessid);
 
-        $message = get_string('deletecheckfull', '', get_string('session', 'attendance'));
+        if ($att->pageparams->action == mod_attendance_sessions_page_params::ACTION_CREATE_CAL_EVENTS) {
+            $message = get_string('createcheckcalevents', 'attendance');
+        } else if ($att->pageparams->action == mod_attendance_sessions_page_params::ACTION_DELETE_CAL_EVENTS) {
+            $message = get_string('deletecheckcalevents', 'attendance');
+        }
+
         $message .= html_writer::empty_tag('br');
         foreach ($sessionsinfo as $sessinfo) {
             $message .= html_writer::empty_tag('br');
